tests(account): Refactor tests.py -- split into multiple files
