+++ allauth/account/tests/test_auth_backends.py
@@ -1,0 +1,75 @@
+from __future__ import absolute_import
+
+from django.test.utils import override_settings
+
+from allauth.account import app_settings
+from allauth.account.auth_backends import AuthenticationBackend
+from allauth.tests import TestCase
+from allauth.utils import get_user_model
+
+
+class AuthenticationBackendTests(TestCase):
+    def setUp(self):
+        user = get_user_model().objects.create(
+            is_active=True, email="john@example.com", username="john"
+        )
+        user.set_password(user.username)
+        user.save()
+        self.user = user
+
+    @override_settings(
+        ACCOUNT_AUTHENTICATION_METHOD=app_settings.AuthenticationMethod.USERNAME
+    )  # noqa
+    def test_auth_by_username(self):
+        user = self.user
+        backend = AuthenticationBackend()
+        self.assertEqual(
+            backend.authenticate(
+                request=None, username=user.username, password=user.username
+            ).pk,
+            user.pk,
+        )
+        self.assertEqual(
+            backend.authenticate(
+                request=None, username=user.email, password=user.username
+            ),
+            None,
+        )
+
+    @override_settings(
+        ACCOUNT_AUTHENTICATION_METHOD=app_settings.AuthenticationMethod.EMAIL
+    )  # noqa
+    def test_auth_by_email(self):
+        user = self.user
+        backend = AuthenticationBackend()
+        self.assertEqual(
+            backend.authenticate(
+                request=None, username=user.email, password=user.username
+            ).pk,
+            user.pk,
+        )
+        self.assertEqual(
+            backend.authenticate(
+                request=None, username=user.username, password=user.username
+            ),
+            None,
+        )
+
+    @override_settings(
+        ACCOUNT_AUTHENTICATION_METHOD=app_settings.AuthenticationMethod.USERNAME_EMAIL
+    )  # noqa
+    def test_auth_by_username_or_email(self):
+        user = self.user
+        backend = AuthenticationBackend()
+        self.assertEqual(
+            backend.authenticate(
+                request=None, username=user.email, password=user.username
+            ).pk,
+            user.pk,
+        )
+        self.assertEqual(
+            backend.authenticate(
+                request=None, username=user.username, password=user.username
+            ).pk,
+            user.pk,
+        )
